% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{AddEstimates}
\alias{AddEstimates}
\title{AddEstimates}
\usage{
AddEstimates(county_pop)
}
\arguments{
\item{county_pop}{a dataset of population by county from 2000 to 2020}
}
\value{
a population dataset with estimates for 1999 and 1998 added
}
\description{
given population dataset made by GetCountyPop, uses lm to predict populations for 1998 and 1999
}
