% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Dataset.R
\name{Build_Dataset}
\alias{Build_Dataset}
\title{Build_Dataset}
\usage{
Build_Dataset()
}
\value{
returns dataset containing data on counties, GPS, deaths, 2016 income, and population
}
\description{
builds dataset containing data on counties, GPS, deaths, 2016 income, and population
}
