% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{ClusterData}
\alias{ClusterData}
\title{ClusterData}
\usage{
ClusterData(df, n_clusters, ...)
}
\arguments{
\item{df}{the dataframe to be sent--build spcifically for dataset with this package}

\item{n_clusters}{the number of kmeans clusters to use}

\item{...}{additional variables for kmeans}
}
\value{
the clustered dataset
}
\description{
Clusters the dataframe by LAT and LONG features
}
