% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{CountBadYears}
\alias{CountBadYears}
\title{CountBadYears}
\usage{
CountBadYears(YEAR, Z_U_RATE, Threshold = 1)
}
\arguments{
\item{YEAR}{the YEAR column from the data}

\item{Z_U_RATE}{zScore of employment data--would work for other columns too}

\item{Threshold}{The number of standard deviations necessary for a year to count as bad.  Defaults to 1}
}
\description{
Counts number of bad yeras based on a given threshold
}
