% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{ExcelToTSV}
\alias{ExcelToTSV}
\title{ExcelToTSV}
\usage{
ExcelToTSV(from_path, to_path, ...)
}
\arguments{
\item{from_path}{address of folder from which to take data--must ONLY include excel files}

\item{to_path}{address of folder in which to write TSV files}

\item{...}{additional params for the read_excel call}
}
\description{
given a folder of excel files, writes a series of tsv files to another given folder
}
