% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Functions.R
\name{GetCountyDeaths}
\alias{GetCountyDeaths}
\title{GetCountyDeaths}
\usage{
GetCountyDeaths()
}
\value{
Returns cleaned dataset of deaths by county
}
\description{
Retrieves and cleans CDC data on deaths per county from inst/extdata/Deaths_by_County folder
}
