% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Functions.R
\name{GetCountyGPS}
\alias{GetCountyGPS}
\title{GetCountyGPS}
\usage{
GetCountyGPS()
}
\value{
a dataset with GPS coordinates for county-FIPS
}
\description{
returns a dataset of county-FIPS codes wtih GPS coordinates.
}
