% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Functions.R
\name{GetCountyPop}
\alias{GetCountyPop}
\title{GetCountyPop}
\usage{
GetCountyPop()
}
\value{
a cleaned dataset of populations by county
}
\description{
returns a cleaned dataset of populations by county from 2000 to 2018
}
