% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Functions.R
\name{GetCountyUnemployment}
\alias{GetCountyUnemployment}
\title{GetCountyUnemployment}
\usage{
GetCountyUnemployment()
}
\value{
a cleaned dataframe of unemployment data
}
\description{
Retrives census figures for unemployment data /inst/extdata/Unemployment_tsv
}
