% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Functions.R
\name{GetWorkingAgePop}
\alias{GetWorkingAgePop}
\title{GetWorkingAgePop}
\usage{
GetWorkingAgePop()
}
\value{
dataframe of county population data by age from 2010 to 2018
}
\description{
retrieves sum total of population by county and year aged 20 to 64
}
