% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{GrowthEstimatesInner}
\alias{GrowthEstimatesInner}
\title{GrowthEstimatesInner}
\usage{
GrowthEstimatesInner(startPop, endPop, ID, col, start = 2000, end = 2010)
}
\arguments{
\item{startPop}{population for starting year}

\item{endPop}{population for ending year}

\item{ID}{the for the county being estimated}

\item{col}{the name of the column being estimated}

\item{start}{starting year}

\item{end}{ending year}
}
\value{
a dataframe of population estimates with features ID, YEAR and column name
}
\description{
Helper function for GrowthEstimates; returns estimates for a single columb given population and years
}
