% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{ilr_partial}
\alias{ilr_partial}
\title{ilr_partial}
\usage{
ilr_partial(df, exceptions)
}
\arguments{
\item{df}{the data to transform}

\item{exceptions}{the columns not to include in the transformation}
}
\value{
the df, with several of the variables transformed, and the exceptions unchanged
}
\description{
performs an ilr transformation on the dataset, but excluding several columns
}
