% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{NaByCol}
\alias{NaByCol}
\title{NaByCol}
\usage{
NaByCol(data)
}
\arguments{
\item{data}{a dataset to print out NAs for}
}
\description{
given a dataset, prints out the number of NAs in each column
}
