% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{predict_logratio}
\alias{predict_logratio}
\title{predict_logratio}
\usage{
predict_logratio(df, model, shift = 0.5, type = -3)
}
\arguments{
\item{df}{the data used to create the model before logratio was applied}

\item{model}{the model using logratios to make predictions on}

\item{shift}{how much of the population to shift from one category to another}

\item{type}{a number of value -3, -2, -1, 1, 2, or 3 that tells the function how to shift the population.  Use value 0 to check}
}
\value{
the change in the target variable--homicides, suicides, or unintentoinal deaths--from applying the shift
}
\description{
given the dataset, shifts some of the population, and then performs a logratio transformation
}
