% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{PredictDeaths}
\alias{PredictDeaths}
\title{PredictDeaths}
\usage{
PredictDeaths(model, data, change = 0.01)
}
\arguments{
\item{model}{the model being used to make prediction}

\item{data}{the dataset being used--cannot differ much from package dataset}

\item{change}{what share of the employed to move to unemployed}
}
\value{
returns predicted number of deaths
}
\description{
gives predictions of number of deaths from model
}
