% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Functions.R
\name{Reconcile}
\alias{Reconcile}
\title{Reconcile}
\usage{
Reconcile(target, goal, circumstances)
}
\arguments{
\item{target}{the feature being reconciled}

\item{goal}{the population that target and other features should sum to}

\item{circumstances}{the original sum of the target and other featuers}
}
\value{
the reconciled population figure
}
\description{
estimates at different levels are not fully commensurate.  This is for helping to reconcile population counts
}
